requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
dateparser==1.1.1
dnspython==2.2.1
firebase-admin==5.2.0
frozenlist==1.3.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
outcome==1.2.0
pandas==1.4.3
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.2.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.10.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.11
websockets==10.3
wsproto==1.1.0
yarl==1.8.1
